- remove deprecated sign up components - rename update-password-form-input
